Remove erroneous while loop
Got rid of a while loop that was in run. that loop would cause a socket to open after closing one, except it wouldn't be part of the data structure, which means it couldn't be closed or accessed or anything really. Removing that prevents extra sockets from being made. I still don't know how to prevent the "socket closed" exception that gets thrown every time the socket is closed. Something to do with the while true loop on 88. I tried using !sock.isClosed() but that didn't help none. Also, got rid of the commented out area.